moved a function to utils file
