+++ src/components/utils.ts
@@ -1,3 +1,5 @@
+import { selector } from "starknet";
+
 export const isUrlFromNethermindDomain = (url: string) => {
   try {
     // Parse the URL using the URL constructor
@@ -44,4 +46,116 @@ export const toCamelCase = (str: string) => {
   return str.replace(/_([a-z])/g, function (match, letter) {
     return letter.toUpperCase();
   });
+};
+
+
+export  const transformParamsToArray = (params: any) => {
+  const transformParam = (param: any): any => {
+    if (param.description || param.placeholder) {
+      if (param.oneOf) {
+        return {
+          description: param.description,
+          index: param.index,
+          value: param.oneOf.map((option: any) => ({
+            name: option.name,
+            placeholder: option.placeholder,
+            pattern: option.pattern,
+            enum: option.enum,
+          })),
+        };
+      } else {
+        return {
+          description: param.description,
+          placeholder: param.placeholder,
+        };
+      }
+    } else if (!param.placeholder && !param.description) {
+      let params = {};
+      for (const [key, value] of Object.entries(param)) {
+        params = {
+          ...params,
+          [key]: {
+            ...transformParam(value),
+            name: key,
+          },
+        };
+      }
+
+      return params;
+    }
+    return {};
+  };
+
+  return params
+    ? Object.entries(params).flatMap(([name, value]) => {
+        if (Array.isArray(value)) {
+          let values = value.map((param: any) => {
+            return transformParam(param);
+          });
+
+          return { name, value: values };
+        }
+        return { name, value: transformParam(value) };
+      })
+    : [];
+};
+
+interface ParamsObject {
+  [key: string]: any;
+}
+
+export const constructParamsArray = (latestParamsArray: any) => {
+  if (latestParamsArray.length === 0) {
+    return [];
+  }
+
+  const processPlaceholder = (latestParams: any, isEntryPoint: boolean) => {
+    let placeholder = latestParams.placeholder;
+    if (isEntryPoint) {
+      placeholder = selector.getSelectorFromName(placeholder);
+    }
+    return placeholder;
+  };
+
+  const constructParams = (latestParams: any, isEntryPoint: boolean): any => {
+    if (!latestParams.description) {
+      return Object.entries(latestParams).reduce(
+        (acc: ParamsObject, [key, value]) => {
+          acc[key] = Array.isArray(value)
+            ? value.map((val) =>
+                constructParams(val, key === "entry_point_selector")
+              )
+            : constructParams(value, key === "entry_point_selector");
+          return acc;
+        },
+        {}
+      );
+    } else if (latestParams?.value?.length > 0) {
+      const selectedOption = latestParams.value[latestParams.index];
+      const placeholder = processPlaceholder(selectedOption, isEntryPoint);
+      return selectedOption.enum
+        ? placeholder
+        : { [selectedOption.name]: placeholder };
+    } else {
+      return processPlaceholder(latestParams, isEntryPoint);
+    }
+  };
+
+  return latestParamsArray.reduce(
+    (accumulator: any, param: { value: any[]; name: string }) => {
+      if (Array.isArray(param.value)) {
+        accumulator[param.name] = param.value.map((val: any) =>
+          constructParams(val, val.name === "entry_point_selector")
+        );
+      } else {
+        accumulator[param.name] = constructParams(
+          param.value,
+          param.name === "entry_point_selector"
+        );
+      }
+
+      return accumulator;
+    },
+    {}
+  );
 };
